add namespace; using size_t;

+++ algorithm/discretization.hpp
@@ -11,7 +11,7 @@ struct discretize {
         sort(vec.begin(), vec.end(), cmp());
         vec.resize(unique(vec.begin(), vec.end()) - vec.begin());
     }
-    ull operator[](const T &x) {
+    size_t operator[](const T &x) {
         return lower_bound(vec.begin(), vec.end(), x, cmp()) - vec.begin();
     }
 };
